Fix Scala 2.11 compilation
